requirements.txt
altgraph==0.17.5
apturl==0.5.2
attrs==18.2.0
autograd==1.8.0
autograd-gamma==0.5.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2024.12.14
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==42.0.1
cupshelpers==1.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
data-wrangler==0.0.3
DataWrangler==0.2
dbus-python==1.2.18
defer==1.0.6
dell-recovery==0.0.0
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.2
docx==0.2.4
duplicity==0.8.21
ecos==2.0.14
et-xmlfile==1.1.0
evdev==1.4.0
exception==0.1.0
exceptiongroup==1.2.2
fasteners==0.14.1
feedparser==6.0.12
fillpdf==0.7.3
fonttools==4.48.1
formulaic==1.2.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.17.1
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.62.0
gpg==1.10.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imagetk==0.1.2
IMAPClient==3.0.1
importlib-metadata==4.6.4
interface-meta==1.3.0
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifelines==0.30.0
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.2
menulibre==2.2.2
mercadopago==2.2.3
monotonic==1.6
more-itertools==8.10.0
mplcyberpunk==0.7.1
narwhals==2.12.0
netifaces==0.11.0
numexpr==2.14.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
osqp==0.6.7.post3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
paramiko==2.9.3
pdf2image==1.17.0
pdflatex==0.1.3
pdfminer.six==20231228
pdfrw2==0.5.0
pefile==2021.9.3
pexpect==4.8.0
pillow==10.4.0
progressbar==2.5
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pyaes==1.6.1
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyDrive2==1.19.0
PyGithub==2.3.0
PyGObject==3.42.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyJWT==2.8.0
pymacaroons==0.13.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pyparsing==3.3.2
pypdf==6.9.1
PyPDF2==3.0.1
pyperclip==1.9.0
pyRFC3339==1.1
PySocks==1.7.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.0
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmail36==1.0.5
qdldl==0.1.7.post5
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.7.2
scikit-survival==0.25.0
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.27.1
setproctitle==1.2.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.10.2
soupsieve==2.3.1
systemd-python==234
tabula-py==2.9.3
Telethon==1.38.1
threadpoolctl==3.6.0
tk==0.1.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xtkinter==1.0.6
zipp==1.0.0
